chore: apply @claude changes for PR #10329
