+++ src/cli/src/format-runtime/index.ts
@@ -1,2 +1,1 @@
 export * from "./entry-point.js";
-export * from "./ignore-path-registry.js";
