+++ src/cli/src/modules/formatting/index.ts
@@ -12,5 +12,5 @@ export {
 export * from "./cache-eviction-policy.js";
 export * from "./format-memory-constants.js";
 export * from "./format-memory-options.js";
-export * from "./ignore-rules-negation-tracker.js";
+export * from "./ignore-tracking.js";
 export * from "./target-path-resolution.js";
